+++ browser-switch/src/test/java/com/braintreepayments/api/BrowserSwitchClientUnitTest.java
@@ -8,6 +8,7 @@ import static org.junit.Assert.assertSame;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.ArgumentMatchers.same;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
@@ -21,6 +22,7 @@ import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
 
+import androidx.activity.ComponentActivity;
 import androidx.fragment.app.FragmentActivity;
 
 import org.json.JSONObject;
@@ -51,6 +53,8 @@ public class BrowserSwitchClientUnitTest {
     private ActivityController<FragmentActivity> activityController;
     private FragmentActivity activity;
 
+    private ActivityController<ComponentActivity> componentActivityController;
+    private ComponentActivity componentActivity;
     @Before
     public void beforeEach() {
         persistentStore = mock(BrowserSwitchPersistentStore.class);
@@ -63,16 +67,17 @@ public class BrowserSwitchClientUnitTest {
         activityController = Robolectric.buildActivity(FragmentActivity.class).setup();
         activity = spy(activityController.get());
 
+        componentActivityController = Robolectric.buildActivity(ComponentActivity.class).setup();
+        componentActivity = spy(componentActivityController.get());
+
         applicationContext = activity.getApplicationContext();
     }
 
     @Test
     public void start_whenActivityIsFinishing_throwsException() {
-        when(browserSwitchInspector.deviceHasBrowser(applicationContext)).thenReturn(true);
         when(browserSwitchInspector.isDeviceConfiguredForDeepLinking(applicationContext, "return-url-scheme")).thenReturn(true);
-        when(browserSwitchInspector.deviceHasChromeCustomTabs(applicationContext)).thenReturn(true);
 
-        when(activity.isFinishing()).thenReturn(true);
+        when(componentActivity.isFinishing()).thenReturn(true);
 
         BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
 
@@ -83,19 +88,14 @@ public class BrowserSwitchClientUnitTest {
                 .returnUrlScheme("return-url-scheme")
                 .metadata(metadata);
 
-        try {
-            sut.start(activity, options);
-            fail("should fail");
-        } catch (BrowserSwitchException e) {
-            assertEquals(e.getMessage(), "Unable to start browser switch while host Activity is finishing.");
-        }
+        BrowserSwitchPendingRequest request = sut.start(componentActivity, options);
+        assertTrue(request instanceof BrowserSwitchPendingRequest.Failure);
+        assertEquals(((BrowserSwitchPendingRequest.Failure) request).getCause().getMessage(), "Unable to start browser switch while host Activity is finishing.");
     }
 
     @Test
-    public void start_whenChromeCustomTabsSupported_createsBrowserSwitchIntentAndInitiatesBrowserSwitch() throws BrowserSwitchException {
-        when(browserSwitchInspector.deviceHasBrowser(applicationContext)).thenReturn(true);
+    public void start_whenSuccessful_returnsBrowserSwitchRequest() {
         when(browserSwitchInspector.isDeviceConfiguredForDeepLinking(applicationContext, "return-url-scheme")).thenReturn(true);
-        when(browserSwitchInspector.deviceHasChromeCustomTabs(applicationContext)).thenReturn(true);
 
         BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
 
@@ -105,16 +105,14 @@ public class BrowserSwitchClientUnitTest {
                 .url(browserSwitchDestinationUrl)
                 .returnUrlScheme("return-url-scheme")
                 .metadata(metadata);
-        sut.start(activity, options);
+        BrowserSwitchPendingRequest browserSwitchPendingRequest = sut.start(componentActivity, options);
 
-        verify(customTabsInternalClient).launchUrl(activity, browserSwitchDestinationUrl, false);
-        verify(activity, never()).startActivity(any(Intent.class));
+        verify(customTabsInternalClient).launchUrl(componentActivity, browserSwitchDestinationUrl, false);
 
-        ArgumentCaptor<BrowserSwitchRequest> captor =
-                ArgumentCaptor.forClass(BrowserSwitchRequest.class);
-        verify(persistentStore).putActiveRequest(captor.capture(), same(applicationContext));
+        assertNotNull(browserSwitchPendingRequest);
+        assertTrue(browserSwitchPendingRequest instanceof BrowserSwitchPendingRequest.Started);
 
-        BrowserSwitchRequest browserSwitchRequest = captor.getValue();
+        BrowserSwitchRequest browserSwitchRequest = ((BrowserSwitchPendingRequest.Started) browserSwitchPendingRequest).getBrowserSwitchRequest();
         assertEquals(browserSwitchRequest.getRequestCode(), 123);
         assertEquals(browserSwitchRequest.getUrl(), browserSwitchDestinationUrl);
         assertSame(browserSwitchRequest.getMetadata(), metadata);
@@ -122,10 +120,9 @@ public class BrowserSwitchClientUnitTest {
     }
 
     @Test
-    public void start_whenChromeCustomTabsNotSupported_createsBrowserSwitchIntentAndInitiatesBrowserSwitch() throws BrowserSwitchException {
-        when(browserSwitchInspector.deviceHasBrowser(applicationContext)).thenReturn(true);
+    public void start_whenNoBrowserAvailable_returnsFailure() {
         when(browserSwitchInspector.isDeviceConfiguredForDeepLinking(applicationContext, "return-url-scheme")).thenReturn(true);
-        when(browserSwitchInspector.deviceHasChromeCustomTabs(applicationContext)).thenReturn(false);
+        doThrow(new ActivityNotFoundException()).when(customTabsInternalClient).launchUrl(any(Context.class), any(Uri.class), eq(false));
 
         BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
 
@@ -135,53 +132,12 @@ public class BrowserSwitchClientUnitTest {
                 .url(browserSwitchDestinationUrl)
                 .returnUrlScheme("return-url-scheme")
                 .metadata(metadata);
-        sut.start(activity, options);
-
-        verify(customTabsInternalClient, never()).launchUrl(activity, browserSwitchDestinationUrl, false);
-
-        ArgumentCaptor<Intent> intentCaptor = ArgumentCaptor.forClass(Intent.class);
-        verify(activity).startActivity(intentCaptor.capture());
-
-        Intent intent = intentCaptor.getValue();
-        assertEquals(intent.getData(), browserSwitchDestinationUrl);
-
-        ArgumentCaptor<BrowserSwitchRequest> captor =
-                ArgumentCaptor.forClass(BrowserSwitchRequest.class);
-        verify(persistentStore).putActiveRequest(captor.capture(), same(applicationContext));
-
-        BrowserSwitchRequest browserSwitchRequest = captor.getValue();
-        assertEquals(browserSwitchRequest.getRequestCode(), 123);
-        assertEquals(browserSwitchRequest.getUrl(), browserSwitchDestinationUrl);
-        assertSame(browserSwitchRequest.getMetadata(), metadata);
-        assertTrue(browserSwitchRequest.getShouldNotifyCancellation());
+        BrowserSwitchPendingRequest request = sut.start(componentActivity, options);
+        assertTrue(request instanceof BrowserSwitchPendingRequest.Failure);
+        assertEquals(((BrowserSwitchPendingRequest.Failure) request).getCause().getMessage(), "Unable to start browser switch without a web browser.");
     }
-
     @Test
-    public void start_whenChromeCustomTabsNotSupported_whenActivityNotFound_throwsBrowserSwitchException() {
-        when(browserSwitchInspector.isDeviceConfiguredForDeepLinking(applicationContext, "return-url-scheme")).thenReturn(true);
-        when(browserSwitchInspector.deviceHasChromeCustomTabs(applicationContext)).thenReturn(false);
-
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
-        doThrow(new ActivityNotFoundException("No browser")).when(activity).startActivity(any(Intent.class));
-
-        JSONObject metadata = new JSONObject();
-        BrowserSwitchOptions options = new BrowserSwitchOptions()
-                .requestCode(123)
-                .url(browserSwitchDestinationUrl)
-                .returnUrlScheme("return-url-scheme")
-                .metadata(metadata);
-
-        try {
-            sut.start(activity, options);
-            fail("should fail");
-        } catch (BrowserSwitchException e) {
-            assertEquals(e.getMessage(), "Unable to start browser switch without a web browser.");
-        }
-    }
-
-    @Test
-    public void start_whenRequestCodeIsIntegerMinValue_throwsError() {
-        when(browserSwitchInspector.deviceHasBrowser(applicationContext)).thenReturn(true);
+    public void start_whenRequestCodeIsIntegerMinValue_returnsFailure() {
         when(browserSwitchInspector.isDeviceConfiguredForDeepLinking(applicationContext, "return-url-scheme")).thenReturn(true);
 
         BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
@@ -192,17 +148,13 @@ public class BrowserSwitchClientUnitTest {
                 .url(browserSwitchDestinationUrl)
                 .returnUrlScheme("return-url-scheme")
                 .metadata(metadata);
-        try {
-            sut.start(activity, options);
-            fail("should fail");
-        } catch (BrowserSwitchException e) {
-            assertEquals(e.getMessage(), "Request code cannot be Integer.MIN_VALUE");
-        }
+        BrowserSwitchPendingRequest request = sut.start(componentActivity, options);
+        assertTrue(request instanceof BrowserSwitchPendingRequest.Failure);
+        assertEquals(((BrowserSwitchPendingRequest.Failure) request).getCause().getMessage(), "Request code cannot be Integer.MIN_VALUE");
     }
 
     @Test
-    public void start_whenDeviceIsNotConfiguredForDeepLinking_throwsError() {
-        when(browserSwitchInspector.deviceHasBrowser(applicationContext)).thenReturn(true);
+    public void start_whenDeviceIsNotConfiguredForDeepLinking_returnsFailure() {
         when(browserSwitchInspector.isDeviceConfiguredForDeepLinking(applicationContext, "return-url-scheme")).thenReturn(false);
 
         BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
@@ -214,20 +166,16 @@ public class BrowserSwitchClientUnitTest {
                 .returnUrlScheme("return-url-scheme")
                 .metadata(metadata);
 
-        try {
-            sut.start(activity, options);
-            fail("should fail");
-        } catch (BrowserSwitchException e) {
-            assertEquals("The return url scheme was not set up, incorrectly set up, or more than one " +
-                    "Activity on this device defines the same url scheme in it's Android Manifest. " +
-                    "See https://github.com/braintree/browser-switch-android for more information on " +
-                    "setting up a return url scheme.", e.getMessage());
-        }
+        BrowserSwitchPendingRequest request = sut.start(componentActivity, options);
+        assertTrue(request instanceof BrowserSwitchPendingRequest.Failure);
+        assertEquals("The return url scheme was not set up, incorrectly set up, or more than one " +
+                "Activity on this device defines the same url scheme in it's Android Manifest. " +
+                "See https://github.com/braintree/browser-switch-android for more information on " +
+                "setting up a return url scheme.", ((BrowserSwitchPendingRequest.Failure) request).getCause().getMessage());
     }
 
     @Test
-    public void start_whenNoReturnUrlSchemeSet_throwsError() {
-        when(browserSwitchInspector.deviceHasBrowser(applicationContext)).thenReturn(true);
+    public void start_whenNoReturnUrlSchemeSet_throwsFailure() {
         when(browserSwitchInspector.isDeviceConfiguredForDeepLinking(applicationContext, "return-url-scheme")).thenReturn(true);
 
         BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
@@ -238,12 +186,9 @@ public class BrowserSwitchClientUnitTest {
                 .returnUrlScheme(null)
                 .url(browserSwitchDestinationUrl)
                 .metadata(metadata);
-        try {
-            sut.start(activity, options);
-            fail("should fail");
-        } catch (BrowserSwitchException e) {
-            assertEquals("A returnUrlScheme is required.", e.getMessage());
-        }
+        BrowserSwitchPendingRequest request = sut.start(componentActivity, options);
+        assertTrue(request instanceof BrowserSwitchPendingRequest.Failure);
+        assertEquals("A returnUrlScheme is required.", ((BrowserSwitchPendingRequest.Failure) request).getCause().getMessage());
     }
 
     @Test
@@ -439,17 +384,16 @@ public class BrowserSwitchClientUnitTest {
     }
 
     @Test
-    public void parseResult_whenActiveRequestMatchesRequestCodeAndDeepLinkResultURLScheme_returnsBrowserSwitchSuccessResult() {
+    public void parseResult_whenActiveRequestMatchesDeepLinkResultURLScheme_returnsBrowserSwitchSuccessResult() {
         BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
 
         JSONObject requestMetadata = new JSONObject();
         BrowserSwitchRequest request =
             new BrowserSwitchRequest(123, browserSwitchDestinationUrl, requestMetadata, "fake-url-scheme", false);
-        when(persistentStore.getActiveRequest(same(applicationContext))).thenReturn(request);
 
         Uri deepLinkUrl = Uri.parse("fake-url-scheme://success");
         Intent intent = new Intent(Intent.ACTION_VIEW, deepLinkUrl);
-        BrowserSwitchResult browserSwitchResult = sut.parseResult(applicationContext, 123, intent);
+        BrowserSwitchResult browserSwitchResult = sut.parseResult(new BrowserSwitchPendingRequest.Started(request), intent);
 
         assertNotNull(browserSwitchResult);
         assertEquals(BrowserSwitchStatus.SUCCESS, browserSwitchResult.getStatus());
@@ -457,45 +401,16 @@ public class BrowserSwitchClientUnitTest {
     }
 
     @Test
-    public void parseResult_whenActiveRequestMatchesRequestCodeAndDeepLinkResultURLSchemeDoesntMatch_returnsNull() {
+    public void parseResult_whenDeepLinkResultURLSchemeDoesntMatch_returnsNull() {
         BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
 
         JSONObject requestMetadata = new JSONObject();
         BrowserSwitchRequest request =
                 new BrowserSwitchRequest(123, browserSwitchDestinationUrl, requestMetadata, "fake-url-scheme", false);
-        when(persistentStore.getActiveRequest(same(applicationContext))).thenReturn(request);
 
         Uri deepLinkUrl = Uri.parse("a-different-url-scheme://success");
         Intent intent = new Intent(Intent.ACTION_VIEW, deepLinkUrl);
-        BrowserSwitchResult browserSwitchResult = sut.parseResult(applicationContext, 123, intent);
-
-        assertNull(browserSwitchResult);
-    }
-
-    @Test
-    public void parseResult_whenActiveRequestDoesntMatchRequestCode_returnsNull() {
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
-
-        JSONObject requestMetadata = new JSONObject();
-        BrowserSwitchRequest request =
-                new BrowserSwitchRequest(456, browserSwitchDestinationUrl, requestMetadata, "fake-url-scheme", false);
-        when(persistentStore.getActiveRequest(same(applicationContext))).thenReturn(request);
-
-        Uri deepLinkUrl = Uri.parse("fake-url-scheme://success");
-        Intent intent = new Intent(Intent.ACTION_VIEW, deepLinkUrl);
-        BrowserSwitchResult browserSwitchResult = sut.parseResult(applicationContext, 123, intent);
-
-        assertNull(browserSwitchResult);
-    }
-
-    @Test
-    public void parseResult_whenNoActiveRequestExists_returnsNull() {
-        BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
-        when(persistentStore.getActiveRequest(same(applicationContext))).thenReturn(null);
-
-        Uri deepLinkUrl = Uri.parse("fake-url-scheme://success");
-        Intent intent = new Intent(Intent.ACTION_VIEW, deepLinkUrl);
-        BrowserSwitchResult browserSwitchResult = sut.parseResult(applicationContext, 123, intent);
+        BrowserSwitchResult browserSwitchResult = sut.parseResult(new BrowserSwitchPendingRequest.Started(request), intent);
 
         assertNull(browserSwitchResult);
     }
@@ -504,7 +419,11 @@ public class BrowserSwitchClientUnitTest {
     public void parseResult_whenIntentIsNull_returnsNull() {
         BrowserSwitchClient sut = new BrowserSwitchClient(browserSwitchInspector, persistentStore, customTabsInternalClient);
 
-        BrowserSwitchResult browserSwitchResult = sut.parseResult(applicationContext, 123, null);
+        JSONObject requestMetadata = new JSONObject();
+        BrowserSwitchRequest request =
+                new BrowserSwitchRequest(123, browserSwitchDestinationUrl, requestMetadata, "fake-url-scheme", false);
+
+        BrowserSwitchResult browserSwitchResult = sut.parseResult(new BrowserSwitchPendingRequest.Started(request), null);
         assertNull(browserSwitchResult);
     }